Utils parseSuite method
