+++ src/com/qaprosoft/Utils.groovy
@@ -1,5 +1,8 @@
 package com.qaprosoft
 
+import org.testng.xml.Parser
+import org.testng.xml.XmlSuite
+
 class Utils {
 
     static def printStackTrace(Exception e) {
@@ -13,4 +16,14 @@ class Utils {
     static def encodeToBase64(stringValue) {
         return stringValue.bytes.encodeBase64().toString()
     }
+
+    static XmlSuite parseSuite(String path) {
+        def xmlFile = new Parser(path)
+        xmlFile.setLoadClasses(false)
+
+        List<XmlSuite> suiteXml = xmlFile.parseToList()
+        XmlSuite currentSuite = suiteXml.get(0)
+        return currentSuite
+    }
+
 }
